Fix runtime type error
The test Table_delete in test_table.cpp failed in one of the test environments.
size_t not supported by the hepler templates.

Add support for 'unsigned'.

+++ test/test_table.cpp
@@ -1029,7 +1029,7 @@ TEST(Table_Delete)
 {
     TestTable01 table;
 
-    for (size_t i = 0; i < 10; ++i) {
+    for (int i = 0; i < 10; ++i) {
         add(table, 0, i, true, Wed);
     }
